i dont wanna talk about it
*proceeds to talk about it*
completely reworked classes to account for upcoming labs

created a component-based physics system where each component applies a force to a velocity vector, and that vector is applied to the transform at the end of the cycle.
currently, the order of force application is dependent on the order within the main class's (PhysicsBody) PhysicsComponent array.
ideally, a priority queue would allow the developer to better select when each force gets applied.
current implemented components are as follows:
gravity,
bounce,
trajectory

all being used by a core component, PhysicsBody.
The body is updated each frame by a PhysicsManager class, in charge of keeping track of physicbodies in the scene