Modified partition system and collisionvolume
collision volume is now an interface, allowing the inheritance of PhysicsComponentBase.

in retrospect, this pattern is less efficient, as it incurs the additional function overhead of storing a get method for the objects transform.

As it stands, however, the system is able to handle each collision interaction. To add another type of collision (i.e. AABB), a new entry must be added to the VolumeType enum, and a new method must be added to ICollisionVolume, essentially increasing the work required to be done in the order of O(N * N-1) (probably)